+++ packages/gateway/src/middleware/index.ts
@@ -1,0 +1,7 @@
+import { RequestMiddleware } from './request-middleware.interface';
+
+const defaultMiddlewares: RequestMiddleware[] = [];
+export * from './forward-headers';
+export * from './load-custom-middlewares';
+export * from './request-middleware.interface';
+export { defaultMiddlewares };