+++ packages/gateway/src/middleware/request-middleware.interface.ts
@@ -1,0 +1,6 @@
+import { ApiDef } from '@graphql-portal/types';
+import { RequestHandler } from 'express';
+
+export interface RequestMiddleware {
+  (apiDef: ApiDef): RequestHandler;
+}